requirements.txt
torch
torchvision
transformers
faiss-cpu
fastapi
uvicorn
pillow